update TagLayout Component
